Changed the generated name of request classes
In order to make the class names proper in terms of English, the
generated names are looking like StoreModelRequest and
UpdateModelRequest now.
Before it was ModelStoreRequest and ModelUpdateRequest.

Still the idea of putting the model name into the request class name
is followed as when the project contains a lot of custom request
classes, it is much easier to maintain them with unique names.